+++ bin/c-mk-delete.js
@@ -1,0 +1,12 @@
+#!/usr/bin/env node
+
+'use strict';
+
+const program = require('commander');
+const { exec } = require('shelljs');
+
+// The basic program, which uses sub-commands.
+program
+    .parse(process.argv);
+
+exec('minikube delete');