+++ bin/c.js
@@ -9,6 +9,8 @@ const { missingCommand } = require('./lib/c');
 program
     .command('d <command>', 'Shortcuts to control Docker.')
     .command('dc <command>', 'Shortcuts to control Docker Compose.')
+    .command('hosts <command>', 'Shortcuts to help with hosts management.')
+    .command('mk <command>', 'Shortcuts to control Minikube.')
     .command('npm <command>', 'Shortcuts to help with NPM.')
     .command('yarn <command>', 'Shortcuts to help with Yarn.')
     .parse(process.argv);