+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 This file is a history of the changes made to @idearium/cli.
 
+## Unreleased
+
+- New commands.
+
+## Improvements.
+
+- Added `c mk delete`, `c mk hosts`, `c mk ip`, `c mk start`, `c mk stop` and `c mk restart` to control Minikube.
+- Added `c hosts add`, `c hosts get` and `c hosts remove` to help with easy hosts management.
+
 ## 1.0.0-alpha.12
 
 - New command.